+++ libs/i18n/src/lib/translations/translations_bg.ts
@@ -604,6 +604,12 @@ export default {
     btpToolHeader: {
         menuButtonAriaLabel: 'Бутон за меню'
     },
+    btpNavigation: {
+        expanderAriaLabel: 'expand/collapse sub-items',
+        moreButtonAriaLabel: 'Displays additional navigation items that are hidden due to limited screen space',
+        overflowMenuAriaLabel: 'Additional Navigation Items',
+        snappedPopoverRoleDescription: 'Navigation List Tree'
+    },
     coreNotification: {
         groupHeaderTitle: 'Разгъване/свиване',
         groupAriaDescription: 'Група за известия',