+++ libs/i18n/src/lib/translations/translations_he.ts
@@ -602,6 +602,12 @@ export default {
     btpToolHeader: {
         menuButtonAriaLabel: 'לחצן תפריט'
     },
+    btpNavigation: {
+        expanderAriaLabel: 'expand/collapse sub-items',
+        moreButtonAriaLabel: 'Displays additional navigation items that are hidden due to limited screen space',
+        overflowMenuAriaLabel: 'Additional Navigation Items',
+        snappedPopoverRoleDescription: 'Navigation List Tree'
+    },
     coreNotification: {
         groupHeaderTitle: 'הרחבה/כיווץ',
         groupAriaDescription: 'קבוצת הודעות',