+++ libs/ui5-webcomponents-fiori/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.58.0-rc.19](https://github.com/SAP/fundamental-ngx/compare/v0.58.0-rc.18...v0.58.0-rc.19) (2025-11-25)
+
+**Note:** Version bump only for package @fundamental-ngx/ui5-webcomponents-fiori
+
+
+
+
+
+# [0.58.0-rc.18](https://github.com/SAP/fundamental-ngx/compare/v0.58.0-rc.17...v0.58.0-rc.18) (2025-11-24)
+
+**Note:** Version bump only for package @fundamental-ngx/ui5-webcomponents-fiori
+
+
+
+
+
+# [0.58.0-rc.17](https://github.com/SAP/fundamental-ngx/compare/v0.58.0-rc.16...v0.58.0-rc.17) (2025-11-21)
+
+**Note:** Version bump only for package @fundamental-ngx/ui5-webcomponents-fiori
+
+
+
+
+
+# [0.58.0-rc.16](https://github.com/SAP/fundamental-ngx/compare/v0.58.0-rc.15...v0.58.0-rc.16) (2025-11-20)
+
+**Note:** Version bump only for package @fundamental-ngx/ui5-webcomponents-fiori
+
+
+
+
+
 # [0.58.0-rc.15](https://github.com/SAP/fundamental-ngx/compare/v0.58.0-rc.14...v0.58.0-rc.15) (2025-11-18)
 
 **Note:** Version bump only for package @fundamental-ngx/ui5-webcomponents-fiori